move some classes
