--- src/main/java/com/cleanroommc/groovyscript/sandbox/FieldAccess.java
@@ -1,13 +1,0 @@
-package com.cleanroommc.groovyscript.sandbox;
-
-public enum FieldAccess {
-    GET, SET;
-
-    public boolean isGetter() {
-        return this == GET;
-    }
-
-    public boolean isSetter() {
-        return this == SET;
-    }
-}
