+++ src/main/java/com/cleanroommc/groovyscript/sandbox/security/GrSMetaClassCreationHandle.java
@@ -1,6 +1,5 @@
-package com.cleanroommc.groovyscript.sandbox.transformer;
+package com.cleanroommc.groovyscript.sandbox.security;
 
-import com.cleanroommc.groovyscript.sandbox.interception.InterceptionManager;
 import groovy.lang.MetaClass;
 import groovy.lang.MetaClassRegistry;
 
@@ -16,7 +15,7 @@ public class GrSMetaClassCreationHandle extends MetaClassRegistry.MetaClassCreat
 
     @Override
     protected MetaClass createNormalMetaClass(Class theClass, MetaClassRegistry registry) {
-        if (!InterceptionManager.INSTANCE.isValid(theClass)) {
+        if (!GroovySecurityManager.INSTANCE.isValid(theClass)) {
             return new BlackListedMetaClass(theClass);
         }
         return super.createNormalMetaClass(theClass, registry);
